% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigma.R
\name{getSigma}
\alias{getSigma}
\title{pairs wil be in the form c("jk", "kr", ...), etc.
e.g. pairs=c("jk", "kr") means we want E(Xj'Xk * Xk'Xr)
with where each of Xj, Xk, Xr is a vector of length p
Assume all X ~ N(0,1) although rows can be correlated,
but columns independent}
\usage{
getSigma(pairs, Diag1 = TRUE, IdentityMatrix = FALSE)
}
\description{
pairs wil be in the form c("jk", "kr", ...), etc.
e.g. pairs=c("jk", "kr") means we want E(Xj'Xk * Xk'Xr)
with where each of Xj, Xk, Xr is a vector of length p
Assume all X ~ N(0,1) although rows can be correlated,
but columns independent
}
