% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WishartMoment.R
\name{WishartMoment}
\alias{WishartMoment}
\title{Function to derive the higher moments of a Wishart distribution symbolically}
\usage{
WishartMoment(input, Diag1 = FALSE, IdentityMatrix = FALSE,
  Sigma = NULL, p = length(input), p.symbol = "p^")
}
\arguments{
\item{input}{For \eqn{E(r_{ab}r_{bc})}{E(r_{ab}*r_{bc})}, for example, type \code{c("ab", "bc")}. See
the details section below. Note that variables must be denoted by single letters.}

\item{Diag1}{Forces \eqn{\Sigma_{11}, \Sigma_{22}}, etc. to be 1.}

\item{IdentityMatrix}{Uses the Identity matrix as \eqn{\Sigma}}

\item{Sigma}{An arbitrary matrix for matrices with constraints. This must be a symmetric
matrix with row and column names corresponding to the variable names. The entries should
be either 0, 1, or a prime number. Equal prime number will constrain the entries to be
the same.}

\item{p}{The parameter \eqn{p} in the Wishart distribution. The result should be the same if \eqn{p} >= \code{length(input)},
although the time taken may be much longer.}

\item{p.symbol}{The name of the parameter to denote the number of observations}
}
\value{
A matrix of counts corresponding to the number of particular terms. For example,
\tabular{rrr}{
 \tab   ac \tab ab*bc \cr
p^2 \tab  0  \tab   1 \cr
p^1 \tab  1 \tab    1
}
means \eqn{\sigma_{ac}p+\sigma_{ab}\sigma_{bc}(p+p^2)}{\sigma_{ac}*p+\sigma_{ac}*\sigma_{bc}*(p+p^2)}.
}
\description{
If \eqn{r_{ab}, r_{bc}, etc.} have a Wishart distribution, this program
calculates \eqn{E(r_{ab}r_{bc}\ldots)}{E(r_{ab}*r_{bc}*\ldots )} symbolically. Currently,
this can only calculate up to the 6th order.
}
\details{
Assume a,b,c are multivariate Normally-distributed vectors with
each element independently having mean 0 and variance \eqn{\sigma_{aa}},
\eqn{\sigma_{bb}},  \eqn{\sigma_{cc}}, respectively.
Assume \deqn{Cov(a_i, b_i) = \sigma_{ab}}
\deqn{Cov(a_i, c_i) = \sigma_{ac},   etc.} and
\deqn{Cov(a_i, b_j) = Cov(a_i, c_j) = 0,   etc. for i != j.}
Let
\deqn{r_{aa} = \sum a_i a_i}
\deqn{r_{ab} = \sum a_i b_i, etc.,}
then
\deqn{r_{aa}, r_{ab}, \ldots \sim Wishart(\Sigma, p)}{r_{aa}, r_{ab}, \ldots ~ Wishart(\Sigma, p)}
where \eqn{\Sigma_{11} = \sigma_{aa}, \Sigma_{12} = \sigma_{ab}, \ldots}.

This program allows you to calculate
\deqn{E(r_{aa}r_{ab}r_{bc}\ldots)}{E(r_{aa}*r_{ab}*r_{bc}\ldots)}
symbolically for arbitrary moments of \eqn{r_{??}}
}
\examples{
WishartMoment(c("ab", "cd"))
WishartMoment(c("ab", "bb", "bc"))

}
